requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
django-htmx==1.17.3
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
telepot==12.7
urllib3==2.2.1
yarl==1.9.4
